split chunk_helper.h
