+++ be/src/column/chunk_helper.h
@@ -1,0 +1,145 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <cstddef>
+#include <deque>
+
+#include "column/vectorized_fwd.h"
+#include "common/global_types.h"
+#include "common/status.h"
+
+namespace starrocks {
+
+// Accumulate small chunk into desired size
+class ChunkAccumulator {
+public:
+    // Avoid accumulate too many chunks in case that chunks' selectivity is very low
+    static inline size_t kAccumulateLimit = 64;
+
+    ChunkAccumulator() = default;
+    ChunkAccumulator(size_t desired_size);
+    void set_desired_size(size_t desired_size);
+    void reset();
+    void finalize();
+    bool empty() const;
+    bool reach_limit() const;
+    Status push(ChunkPtr&& chunk);
+    ChunkPtr pull();
+
+private:
+    size_t _desired_size;
+    ChunkPtr _tmp_chunk;
+    std::deque<ChunkPtr> _output;
+    size_t _accumulate_count = 0;
+};
+
+class ChunkPipelineAccumulator {
+public:
+    ChunkPipelineAccumulator() = default;
+    void set_max_size(size_t max_size) { _max_size = max_size; }
+    void push(const ChunkPtr& chunk);
+    ChunkPtr& pull();
+    void finalize();
+    void reset();
+    void reset_state();
+
+    bool has_output() const;
+    bool need_input() const;
+    bool is_finished() const;
+
+private:
+    static bool _check_json_schema_equallity(const Chunk* one, const Chunk* two);
+
+private:
+    static constexpr double LOW_WATERMARK_ROWS_RATE = 0.75; // 0.75 * chunk_size
+#ifdef BE_TEST
+    static constexpr size_t LOW_WATERMARK_BYTES = 64 * 1024; // 64KB.
+#else
+    static constexpr size_t LOW_WATERMARK_BYTES = 256 * 1024 * 1024; // 256MB.
+#endif
+    ChunkPtr _in_chunk = nullptr;
+    ChunkPtr _out_chunk = nullptr;
+    size_t _max_size = 4096;
+    // For bitmap columns, the cost of calculating mem_usage is relatively high,
+    // so incremental calculation is used to avoid becoming a performance bottleneck.
+    size_t _mem_usage = 0;
+    bool _finalized = false;
+};
+
+class SegmentedColumn final : public std::enable_shared_from_this<SegmentedColumn> {
+public:
+    SegmentedColumn(SegmentedChunkPtr chunk, size_t column_index);
+    SegmentedColumn(std::vector<ColumnPtr> columns, size_t segment_size);
+    ~SegmentedColumn() = default;
+
+    ColumnPtr clone_selective(const uint32_t* indexes, uint32_t from, uint32_t size);
+    ColumnPtr materialize() const;
+
+    bool is_nullable() const;
+    bool has_null() const;
+    size_t size() const;
+    void upgrade_to_nullable();
+    size_t segment_size() const;
+    std::vector<ColumnPtr> columns() const;
+
+private:
+    SegmentedChunkPtr _chunk; // The chunk it belongs to
+    size_t _column_index;     // The index in original chunk
+    const size_t _segment_size;
+
+    std::vector<ColumnPtr> _cached_columns; // Only used for SelectiveCopy
+};
+
+// A big-chunk would be segmented into multi small ones, to avoid allocating large-continuous memory
+// It's not a transparent replacement for Chunk, but must be aware of and set a reasonale chunk_size
+class SegmentedChunk final : public std::enable_shared_from_this<SegmentedChunk> {
+public:
+    SegmentedChunk(size_t segment_size);
+    ~SegmentedChunk() = default;
+
+    static SegmentedChunkPtr create(size_t segment_size);
+
+    void append_column(ColumnPtr column, SlotId slot_id);
+    void append_chunk(const ChunkPtr& chunk, const std::vector<SlotId>& slots);
+    void append_chunk(const ChunkPtr& chunk);
+    void append(const SegmentedChunkPtr& chunk, size_t offset);
+    void build_columns();
+
+    SegmentedColumnPtr get_column_by_slot_id(SlotId slot_id);
+    const SegmentedColumns& columns() const;
+    SegmentedColumns& columns();
+    size_t num_segments() const;
+    const std::vector<ChunkPtr>& segments() const;
+    std::vector<ChunkPtr>& segments();
+    ChunkUniquePtr clone_empty(size_t reserve);
+
+    size_t segment_size() const;
+    void reset();
+    size_t memory_usage() const;
+    size_t num_rows() const;
+    Status upgrade_if_overflow();
+    Status downgrade();
+    bool has_large_column() const;
+    void check_or_die();
+
+private:
+    std::vector<ChunkPtr> _segments;
+    SegmentedColumns _columns;
+
+    const size_t _segment_size;
+};
+
+} // namespace starrocks
