+++ be/src/exec/pipeline/adaptive/collect_stats_context.h
@@ -14,10 +14,10 @@
 
 #pragma once
 
+#include "column/chunk_helper.h"
 #include "column/vectorized_fwd.h"
 #include "exec/pipeline/adaptive/adaptive_fwd.h"
 #include "exec/pipeline/context_with_dependency.h"
-#include "storage/chunk_helper.h"
 #include "util/moodycamel/concurrentqueue.h"
 
 namespace starrocks {
