+++ be/src/exec/pipeline/aggregate/spillable_aggregate_blocking_source_operator.h
@@ -16,11 +16,11 @@
 
 #include <utility>
 
+#include "column/chunk_helper.h"
 #include "exec/aggregator.h"
 #include "exec/pipeline/aggregate/aggregate_blocking_source_operator.h"
 #include "exec/sorted_streaming_aggregator.h"
 #include "runtime/runtime_state.h"
-#include "storage/chunk_helper.h"
 
 namespace starrocks::pipeline {
 class SpillableAggregateBlockingSourceOperator final : public AggregateBlockingSourceOperator {
