+++ be/src/exec/pipeline/sort/partition_sort_sink_operator.cpp
@@ -16,6 +16,7 @@
 
 #include <memory>
 
+#include "column/chunk_helper.h"
 #include "exec/chunks_sorter.h"
 #include "exec/chunks_sorter_full_sort.h"
 #include "exec/chunks_sorter_heap_sort.h"
@@ -30,7 +31,6 @@
 #include "runtime/exec_env.h"
 #include "runtime/runtime_filter_worker.h"
 #include "runtime/runtime_state.h"
-#include "storage/chunk_helper.h"
 #include "types/logical_type.h"
 
 using namespace starrocks;
