+++ be/src/exec/stream/scan/stream_scan_operator.h
@@ -16,6 +16,7 @@
 
 #include <queue>
 
+#include "column/chunk_helper.h"
 #include "common/logging.h"
 #include "connector/connector.h"
 #include "exec/pipeline/fragment_context.h"
@@ -26,7 +27,6 @@
 #include "exec/pipeline/scan/scan_operator.h"
 #include "exec/pipeline/stream_epoch_manager.h"
 #include "exec/workgroup/work_group_fwd.h"
-#include "storage/chunk_helper.h"
 
 namespace starrocks::pipeline {
