+++ be/src/tools/meta_tool.cpp
@@ -60,6 +60,7 @@
 #include "storage/lake/vacuum.h"
 #include "storage/olap_common.h"
 #include "storage/olap_define.h"
+#include "storage/olap_type_infra.h"
 #include "storage/options.h"
 #include "storage/primary_key_dump.h"
 #include "storage/rowset/binary_plain_page.h"
